More QoL stuff
Added an icon for the program while on the taskbar and changed the icon in the top left  of codescans. Newer versions now look for older versions on the desktop and delete them if they're there so that it minimizes the chance of using an older version that has issues.