More updates
Set a variable for console logging so  I don't have to uncomment/recomment every time. Added a Determine Document class because  it was setting doc_type and item_type in the table to lowercase abbreviations (eg. "Purchase Order" was "po"), which was inconsistent with the prior entries.

Also fixed a major issue with the dmsPath variable in Process Uploads. It was set to "//dnas1/dms/documents"  instead  of "../Documents"  which was causing  a different path_id to be generated.