Cleaned up, deleted, added, and/or fixed a ton of stuff.
Cleaned up a lot of unused/commented out code that will no longer be used.

Changed the functioning of the input to accept full work order numbers "000000-000", then it'll find the sales order that goes with it, and rename the file to "XX_XX_#######_######-###_#.pdf" format. This helps to differentiate which work order is being clicked on rather than clicking all of the same sales order number labeled files until they find the correct one.

Fixed document naming issues with VendInfo and No-So.

Commented out and rewrote a whole chunk of ProcessUploads, but didn't delete it because I need to be sure it didn't break anything.

Log4J2 now has a directory on DNAS so that it can easily be checked.